+++ lib/src/component_declaration/builder_helpers.dart
@@ -14,8 +14,6 @@
 
 library over_react.component_declaration.builder_helpers;
 
-import 'package:over_react/src/util/map_util.dart';
-
 import '../../component_base.dart';
 import '../../over_react.dart';
 import './component_base.dart' as component_base;
@@ -134,56 +132,6 @@ abstract class UiProps extends component_base.UiProps with GeneratedClass {
   /// This can be used to derive consumed props by usage in conjunction with [addUnconsumedProps]
   /// and [addUnconsumedDomProps].
   @toBeGenerated PropsMetaCollection get staticMeta => throw UngeneratedError(member: #meta);
-
-  /// Copies key-value pairs from the provided [props] map into this map,
-  /// excluding those with keys found in [consumedProps].
-  ///
-  /// [consumedProps] should be a `Iterable<PropsMeta>` instance.
-  /// This is the return type of [PropsMetaCollection]'s related APIs `forMixins`,
-  /// `allExceptForMixins`, and `all`.
-  ///
-  /// __Example:__
-  ///
-  /// ```dart
-  /// // within a functional component (wrapped in `uiFunction`)
-  /// // Consider props in FooProps "consumed"...
-  /// final consumedProps = props.staticMeta.forMixins({FooProps});
-  /// // ...and filter them out when forwarding props to Bar.
-  /// return (Bar()..addUnconsumedProps(props, consumedProps))();
-  /// ```
-  ///
-  /// To only add DOM props, use [addUnconsumedDomProps].
-  ///
-  /// Related: [UiComponent2]'s `addUnconsumedProps`
-  void addUnconsumedProps(Map props, Iterable<PropsMeta> consumedProps) {
-    final consumedPropKeys = consumedProps.map((consumedProps) => consumedProps.keys);
-    forwardUnconsumedPropsV2(props, propsToUpdate: this, keySetsToOmit: consumedPropKeys);
-  }
-
-  /// Copies DOM only key-value pairs from the provided [props] map into this map,
-  /// excluding those with keys found in [consumedProps].
-  ///
-  /// [consumedProps] should be a `Iterable<PropsMeta>` instance.
-  /// This is the return type of [PropsMetaCollection]'s related APIs `forMixins`,
-  /// `allExceptForMixins`, and `all`.
-  ///
-  /// __Example:__
-  ///
-  /// ```dart
-  /// // within a functional component (wrapped in `uiFunction`)
-  /// // Consider props in FooProps "consumed"...
-  /// final consumedProps = [PropsMeta.forSimpleKey('className')];
-  /// // ...and filter them out when forwarding props to Bar.
-  /// return (Bar()..addUnconsumedDomProps(props, consumedProps))();
-  /// ```
-  ///
-  /// To add all unconsumed props, including DOM props, use [addUnconsumedProps].
-  ///
-  /// Related: [UiComponent2]'s `addUnconsumedDomProps`
-  void addUnconsumedDomProps(Map props, Iterable<PropsMeta> consumedProps) {
-    final consumedPropKeys = consumedProps.map((consumedProps) => consumedProps.keys);
-    forwardUnconsumedPropsV2(props, propsToUpdate: this, keySetsToOmit: consumedPropKeys, onlyCopyDomProps: true);
-  }
 }
 
 /// A [dart.collection.MapView]-like class with strongly-typed getters/setters for React state.
